Move WeaponOptions into separate file and pass in parsed options to weapon search
